+++ src/Raven.CodeAnalysis/Symbols/Constructed/ConstructedNamedTypeSymbol.cs
@@ -939,6 +939,7 @@ internal sealed class ConstructedNamedTypeSymbol : INamedTypeSymbol, IUnionSymbo
     public bool IsAbstract => _originalDefinition.IsAbstract;
     public bool IsClosed => _originalDefinition.IsClosed;
     public bool IsRefLikeType => _originalDefinition.IsRefLikeType;
+    public bool IsReadOnly => _originalDefinition.IsReadOnly;
     public bool IsGenericType => _originalDefinition.IsGenericType;
     public bool IsUnboundGenericType => false;
     public ImmutableArray<INamedTypeSymbol> Interfaces =>